[Fix] #3- Feed Model 폴더로 이동
